+++ snow-schema/src/syntax/scan.ts
