+++ snow-schema/src/syntax/stream/char-stream.ts
