+++ snow-schema/src/syntax/syntax-node.ts
