+++ snow-schema/src/syntax/token.ts
