+++ snow-schema/src/syntax/util/error.ts
